Reorganize measurement sub-package modules
Make the layout a little more conventional by moving the higher
level mgmt components into the root module whilst keeping apps
and helpers in submodules.

More specifically,
- Bring `Measurers` and data set loaders up into the package module
- drop the CDR app into cdr.py
- move `DictProxy` into utils.py
